refactor: 경험 보드 정렬 상태 로직 분리
